requirements.txt
pyyaml>=6.0
argparse>=1.4.0
atlassian-python-api>=3.32.0 